Mock ironic-lib properly in test_deploy_utils
Many of the tests in this file asserted that methods internal to
ironic-lib were called. This has caused gate failures in the past when
ironic-lib changes internal logic.

This mocks out ironic-lib so we only test that the correct methods are
called from ironic's perspective as an external consumer of ironic-lib.

Change-Id: I91d3c7d7b8f4dd6b7652e0cac90ea562ae710636
Partial-Bug: #1552843
Co-Authored-By: John L. Villalovos <john.l.villalovos@intel.com>