Remove 'agent_last_heartbeat' from node.driver_internal_info
node.driver_internal_info['agent_last_heartbeat'] was not being used by
ironic, and Ic198f6a952ed135099bcc525ef548a4aea005056 removed the code
that sets that value.

However, for existing nodes, driver_internal_info['agent_last_heartbeat']
may already exist, so we need to delete it. This patch does that.

In addition, a LOG line was moved (and a nit addressed) to make the code
a bit easier to read.

This is a followup to Ic198f6a952ed135099bcc525ef548a4aea005056.

Change-Id: I194d22c0e605b187823ceebae94d3ef895fb91f5
Related-Bug: #1602410