DevStack: Fix standard PXE on Ubuntu Xenial
The pxelinux.0 file in Ubuntu Xenial is now distributed by the
"pxelinux" package and lives at /usr/lib/PXELINUX/pxelinux.0 instead
of /usr/lib/syslinux/pxelinux.0.

This patch refactors the get_pxe_boot_file() function to accommodate to
be able to find the pxelinux.0 binary in the new path and also creates a
new setup_syslinux_modules() function to place the *.c32 modules in the
TFTP root dir (which in Xenial doesn't live at the same directory as the
PXE image anymore).

Change-Id: I9bd412f10f2235f1979a12dd3004fc60d4aa922f
Closes-Bug: #1638940