Simplify base interfaces in ironic.drivers.base
Move abstract definitions of get_properties and validate to BaseInterface
and make all interfaces inherit it. This has a side effect of them getting
clean steps methods, but they anyway won't be wired in due to a whitelist
of interfaces in the conductor.

The RAIDInterface contained a mistake: it didn't have an ABCMeta metaclass,
so its @abstractmethod definitions didn't work. I've implemented get_properties
for it to avoid breaking implementers.

Also added interface_type to all interfaces for consistency and for future
use in the driver composition work.

Change-Id: Ia6708247a99ecdcf73d83e66350e8c309090b7d4
Partial-Bug: #1524745