Add PS4 for better logfile information of devstack runs
The scripts
    tools/ironic/scripts/cleanup-node.sh
    tools/ironic/scripts/create-node.sh
    tools/ironic/scripts/setup-network.sh

are executed by sudo and the PS4 value disappears during sudo. Add a
PS4 value so that we get better logs during their execution.

PS4 is used to prefix the tracing output when using:
    'set -x' or 'set -o xtrace'

From the bash manpage:
PS4     The  value  of this parameter is expanded as with PS1 and the
        value is printed before each command bash displays during an
        execution trace.  The first character of PS4 is replicated
        multiple times, as necessary, to indicate multiple levels of
        indirection.  The default is ``+ ''.

Change-Id: If06f079bed16d9c84569594aa64259bf06f91a1e
(cherry picked from commit a78097bf3fcbea81fc00cab10a76c709d2d3a1c2)