Add node (database and objects) fields for all interfaces
For the driver composition reform we need nodes to record hardware
interfaces they're using. This change adds necessary fields.

The new fields will have a default value of None, which is the expected
value for nodes using classic drivers.

The new fields are not added to notifications yet, as they're not wired in.

Change-Id: Id7697de0276e9b2730ea62f9c562ed6e1d0f8a06
Partial-Bug: #1524745