Ensures that OneView nodes are free for use by Ironic
Following the requirement. The validation is used within
"OneViewPower.validate". As a result it enforces that a machine
is owned by OneView, thus it cannot be powered on or off by the
Ironic driver. Raising an exception when the node is owned by
Oneview.

Conflicts:
	ironic/drivers/modules/oneview/power.py
	ironic/tests/unit/drivers/modules/oneview/test_deploy_utils.py

Change-Id: Ibf500577b7a3b5976fd49fd217c1be3d3e9c0f46
Closes-Bug: 1627034
Co-Authored-By: Stenio Araujo <steniaraujo@lsd.ufcg.edu.br>
(cherry picked from commit 2505cb58f74b05c534a6a513080919063cc65fa9)