Correct DB Interface migration test
The DB interface migration test laned as part of
Id7697de0276e9b2730ea62f9c562ed6e1d0f8a06 obtained
the list of expected interfaces after the migration
from the current state of the code.

This effectively blocks future interfaces from being
added as the later database migration to create the
interface will not yet have executed when the test
is executed.

As a result, the test has been updated to utilize
a static list of interfaces types, matching those
added as part of the database migration.

Change-Id: I05121ec893f94ccc2fea6ed11992214b303c8367