Update help for 'provisioning_network' option
This updates the help string for [neutron]provisioning_network
configuration option to indicate that the value can also be a
network name.

The release note for this new configuration option is also
updated.

This is a follow up to b6a2919b6198c8fb24d6f7e597f79a08f6ffe049.

Change-Id: I5773b20c5578fa4d14f80de9119e442edfe98dd0