Switch to new location for oslo.db test cases
Oslo.db moved some of their test cases around, they left deprecation
warnings on the old locations which needed resolving. This patch
switches our migration tests to use the new locations and get rid of the
warnings.

Also it removes the test_connect_fail tests, as they use the values that
are deprecated on the oslo_db side.

Change-Id: Idfab443dba26e94d9e50814cf7ffc5624748b476
Closes-Bug: #1647733
Closes-Bug: #1647735