Rolling upgrades related dev documentation
Adds a new 'Rolling Upgrades' developer document that describes the
design (from the associated spec) and includes notes on what to do
when developing a new feature or modifying an IronicObject.

Replaces the rolling upgrades information in the code contribution
guide, with a link to this new Rolling Upgrades page.

Change-Id: I5cc0b42da6e9e66f5c607bddbb3a2232cda97c6f
Closes-Bug: #1526283