Add the new capabilities to the iLO InspectInterface
This patch adds the new capabilities to iLO InspectInterface
as listed in the specification. It requires proliantutils
version >=2.4.0 for these new capabilities.

This patch also removes three capabilities keys from
iLO InspectInterface - ``max_raid_level``, ``BootMode``, and
``sr_iov_devices``.
max_raid_level - It cannot be discovered out-of-band using
proliantutils. The capability is set during RAID configuration
as ``raid_level`` instead.
sr_iov_devices - It cannot be discovered out-of-band. Required
Inband mechanism to discover this.
BootMode - This capability required Nova to change as the value
was supposed to be a list of supported boot modes on the system.
Nova asked to implement the Boolean capabilities instead.

Partial-Bug: 1599425

Change-Id: I64c3e471d55c2020204cf7a20e50c399e43eccf7