Don't override device_owner for tenant network ports
When a vif is passed to us from nova as a tenant port we shouldn't
change the device_owner or device_id because that is what links the port
to the nova instance. This enables the neutron nova notifier to trigger
the correct events in nova for when the neutron port changes, e.g. being
deleted, triggers the detach interface endpoint.

Change-Id: I43c3af9f424a65211ef5a39f13e4810072997339
Closes-Bug: #1656010