Use delay configoption for ssh.SSHPower drivers
In Newton (6.0.0), the configuration option
'[neutron]/port_setup_delay' was introduced [1].
In order not to break ssh power drivers, if that
option was set to 0 (default value), we coded it
to be 15 seconds.

The deprecation period for that is over; this
removes that code. Users should explicitly set the
option to the desired value.

[1] https://review.openstack.org/#/c/293876/

Change-Id: Iffddb1087dcb424101bf693d4c2ae83e1639fd17
Fixes-Bug: #1661050