Fix usage of various deprecated methods
Our unittest output complains about some of the codebase,
described in more detail in the referred bug. This change
fixes all of those places.

static_root is now set only if debug=True is set in pecan
configuration.

Logging is now setup as part of the conf module import,
so that unittests can use it.

Closes-Bug: #1668240
Change-Id: I2536bd2b174e0d224ddf6806f1be37bd29017669