Pin version of virtualbmc in devstack plugin
Virtualbmc is absent from global requirements and upper constraints,
but it depends on pyghmi which is in those requirements.

With new releases of virtualbmc it might lead to newer version of
virtualbmc installed together with old, unsupported by it version of
pyghmi, breaking vbmc functionality.

Until virtualbmc is introduced into g-r and u-c, let's pin the version
of installed virtualbmc to the latest one that existed at the time when
the stable branch was created (0.0.4 for stable/mitaka).

Change-Id: Ic964f0dda043f927547962dd8fd78cb11b4f7064