Properly allow Ironic headers in REST API
In 7fd26bf8031041a0e6df261c71101e29e8bf9610 ("Fix usage of various deprecated
methods", 2017-02-27), the `app.set_latent()` call was replaced with
`cors_middleware.set_defaults()`.  While the mentioned deprecation warning
suggests that the `set_defaults()` function replaces the former `set_latent()`,
they accomplish the same goal but with a different implementation.

`set_defaults()` allows you to specify certain ready-made configration options
in case the developer doesn't provide them in ironic.conf.  However,
`set_latent()` provides configuration that is additive in nature.  This means
that values that have been set with `set_defaults()` will be overridden by
values in ironic.conf, while those set by `set_latent()` will not.

This becomes a problem in the case of the Ironic-specific headers, e.g.
`X-OpenStack-Ironic-API-Version`.  In the Ironic API configuration, we need
explicitely allow these headers in incoming requests.  Otherwise, the request
will be rejected on the grounds that a foreign header is present.  Using
`set_defaults()` to accomplish this doesn't work because those Ironic-specific
headers will be replaced by whatever is set in ironic.conf.

The solution is to add these headers to the `simple_headers` list which is used
as a base when constructing the final list of allowed headers in incoming
OPTIONS requests.  We subclass the real `CORS` class from oslo.middleware and
add the headers to the list.

Change-Id: Ib2b6a11be276220e31d305f4b0a87843d5418958