Raise HTTP 400 rather than 500 error
Currently conductor method inspecting hardware
raises HardwareInspectionFailure with 500 Internal Error
if driver.power.validate fail (e.g. , ``driver_info`` is
not provided or some fields are missing).

Since this is an apparent client error, an HTTP error code
400-Bad request is more appropriate.
The validation method actually raises this needed error
and catching this is not needed anymore.

Change-Id: I080dedeac7ce33135fde8c53494e618ccf07c941
Closes-Bug: #1686457