Update reno for new ilo hardware type
This updates the release note for the new 'ilo' hardware type
to mention that it also supports 'iscsi' deploy interface.

The separate release note about ilo supporting iscsi deploy interface
is removed.

This is all going into the same release, and it is confusing to have
two separate release notes about the ilo hardware type.

Change-Id: I4087d7f805dc7710b554c102f428b92cead1ebf9
Related-Bug: #1689274