Prevent tests from using utils.execute()
This change modifies the base test class so that it mocks out
utils.execute() and similar functions and forces an exception if it
gets called. If utils.execute() is mocked by the test case then it
will work. What this does is prevent un-mocked access to
utils.execute() / processutils.execute() and similar subprocess
library functions.

Inspired by Julian Edwards' patch to ironic-python-agent

Change-Id: Id3a8e6c5fce44a1f5f6f936d2a35e0660adbf086