python-oneviewclient version bump
Updating the minimum version of python-oneviewclient in stable/newton.

We noticed that the CI server have been using the latest version of python-oneviewclient and
altered our jobs to use the minimum version specified at driver-requirements.
When attempting to run stable/newton with python-oneviewclient==2.0.2 we noticed
that the driver never worked with this version.

The minimum required version for stable/newton is now 2.5.1.

Closes-Bug: 1694297
Change-Id: Id396e529de3eba306421dac91e205f4b6a52573f