Support setting inbound global-request-id
As ironic doesn't use oslo.middleware for accepting the inbound
x-openstack-request-id header, this produces the same result as
oslo.middleware does, including the enforcement that it's well formed,
and sets it into the same bit of request environment that will make it
possible for oslo.context to extract it.

This would allow Nova to start calling Ironic with it's request id,
which would show up in Ironic logs as the global_request_id.

oslo spec I65de8261746b25d45e105394f4eeb95b9cb3bd42

Change-Id: I3bc7dbe5f65f10fcb5c6300c1ba09a7a8e8e4c83