BFV Deploy skip minor logging, logic, and test fixes
This is a follow-up to I62e622a2b053f685c2da42ca5106bdb9bdd22dc6
addressing some minor logging, test fixes/additions that were requested
as well as correcting a minor logic discrepency between iscsi and agent
deploy preparation methods as it makes no sense to re-plug networks in
the iscsi deployment preparation method if we are not writing an image
since we will ultimately skip deployment.

Additionally, as a result of review feedback, the FlatNetwork mocks
were updated to be more consistent.

Change-Id: If6bd7e34ea92add15c3c7d8b94a2efbaf882d6ff
Partial-Bug: #1559691