Update comments related to ipmi & old BMCs
The help string for [ipmi]/command_retry_timeout is incorrect. From Sam Betts [1]:
so looking at when this option was introduced ... it was originally introduced
by the ipminative driver which used it as the timeout for waiting until the power
state changed to the requested state. Reading between the lines in the commits
that added this option and documentation, the "setting this too low might break
things" comment originates from bugs in IPMInative. Later on the IPMItool driver
then overloaded this configuration option and changed its meaning by introducing
the min_command_interval config option. It continued to be used for the timeout
for waiting for requested state, but it also now contributed to the the number
of times a IPMItool command would retry retryable errors, which was/is worked out
by dividing retry_timeout by min_command_interval. Now IPMInative no longer in
tree I think we should clean up some of this information.

[1] https://review.openstack.org/#/c/482631/

Change-Id: I8cd8e25a2fb224d477799a2e561573406f9427a9