grenade: For multi-node grenade, do not upgrade nova
Multi-node grenade jobs will fail if we upgrade nova as the minimum
required API for nova to communicate with ironic was upgraded.

Our goal for multi-node grenade is to test rolling upgrades. The
scenario we envision for rolling upgrades to be done is:

 1) Upgrade ironic-conductor services one by one. This means old
    ironic-api services communicating with new and old
    ironic-conductor services. This is what the multi-node grenade job
    tests.
 2) Upgrade ironic-api services.
 3) Upgrade nova

In this upgrade procedure we should not be running updated nova with
old ironic-api services. This patch fixes that.

Closes-Bug: #1706418
Change-Id: I85c607ae7d031e4cabcb505915aca7e88df166aa