Fix broken links to ocata installation guide
Many links in ocata installation guide is broken due to doc
migration, it's originally reported as broken link to Image
Builders. This patch will try to cover all broken links and
anchors found in the installation guide for ocata branch.

As an example, the wrong link to Image Builders is:

http://docs.openstack.org/developer/ironic-python-agent/ocata/index.html#image-builders

We have to remove 'developer' from the path for link to be available.

Change-Id: Id5cf2b22ae2a81cba0102e2566a258b27a3890f2
Closes-Bug: #1705498