Enable no IP address to be returned
Currently 'noop' network_interface users are unable to use
grub to pxe boot machines as the original code path expected
always use IP addresses for the lookup. We've previously
changed the default template, so now there is really no
reason to error at this point, since we should be able to
continue ahead and still boot the instance.

Change-Id: I928ad2c493802a60d77afdb7d4827d820652512c
Story: 1683777
Task: 33759