Add ability to provide configdrive when rebuilding
Previously, the configdrive could only be set when setting
the node's provisioning state to "active". When rebuilding,
the old configdrive was used and therefore was never updated
with latest content.

This change introduces the API microversion 1.35 which will allow
configdrive to be provided when setting the node's provisioning state
to "rebuild".

Closes-bug: #1575935
Change-Id: I9a5529f9fa796c75621e9f4354886bf3032cc248