Document setting discover_hosts_in_cells_interval in nova.conf
By default, starting with Ocata, an operator needs to run a nova-manage command
after each node is enrolled to make sure it's correctly mapped to a cell.
This is not something we can ask in all cases (especially if Ironic UI is used),
so let us document running a periodic task for it instead.

Change-Id: I0c7bf29b0e27bef7205880ebc164d3a82e7b0be8
Closes-Bug: #1715646