Revert "Change pxe dhcp options name to codes."
This reverts commit c377f5cbbd034e16b68a3fc30e138b03badc9c94.

Rajini Karthik discovered that this broke the hardware builds (dell and
oneview, maybe others) -- it caused a TFTP timeout issue. Builds started
failing after it merged (approximately Sep15th midnight) and the builds
on stable/ocata and stable/pike are passing. Rajini reverted this patch
on the Dell CI temporarily and the tests passed.

Change-Id: Id1ef4441511a36f8fc6d639f4bb1dd52cd630445