Remove 'next' for GET /nodes?limit=1&instance_uuid=
For a "GET /nodes?limit=1&instance_uuid=<uuid>" request that has a
node that matches, the response would include a 'next' link.
(Note that the generated 'next' link happens to be incorrect; it
doesn't include 'instance_uuid'). However, at most one node can match
the specified instance UUID, so we should not be returning any 'next'
link.

This fixes it so no 'next' link is returned.

Change-Id: I4efe06652baf3423790c238d4a8d2bc5a736b3a5
Closes-Bug: #1718683