Cleanup unit tests for ipmittool
By having a mock patch on the testcase class it was requiring 40+
functions to have 'mock_sleep' in their argument list. When we are
only checking the mock about five times. Instead put it into the
setUp() function.

Also fix mistakes introduced by previous patches that did not realize
that:
    @mock.patch.object(ipmi, '_make_password_file', _make_password_file_stub)
    def some_func(self, mock_pass): ...

The mock.patch.object() is defining the object completely and thus it
(mock_pass) is not actually passed as an argument to the function.
When removing the 'mock_sleep' mock I discovered this issue.

Change-Id: Iac5dfa8382d04bead745260598d5228c6114b70d