Reduce complexity of node_power_action() function
While attempting to modify the node_power_action() function it caused
the complexity to exceed the 'pep8' check limits. This patch reduces
the complexity of the node_power_action() function by refactoring it.
This will allow future changes to node_power_action() to be allowed.

Added unit tests to test the functions that were pulled out.

Change-Id: I7f84ff23401997c2b92759c306a7bff0589c7f4c