Reduce complexity of node_power_action() function
While attempting to modify the node_power_action() function it caused
the complexity to exceed the 'pep8' check limits. This patch reduces
the complexity of the node_power_action() function by refactoring it.
This will allow future changes to node_power_action() to be allowed.

Added unit tests to test the functions that were pulled out.

This patch is required to be back-ported in order to back-port the
commit 8ceaad42ff363581efd2f43737456d479ea980d7.

Change-Id: I7f84ff23401997c2b92759c306a7bff0589c7f4c
(cherry picked from commit 0d788525dfa0a7dee11b005318478a7cee597055)