Stop passing raw Exceptions as the reasons for ironic Image exceptions
When an exception occured in the ironic image service code, the raw
exception was passed into the IronicException as the reason. The
ironic exception logic then attempted to converted it to json which
fails and then falls back to converting it to string. This patch
preemptively converts the Exception to string to avoid that first
failure.

Change-Id: Id3a166a74505c142d7217217882246a64f625eb8