Get a new OneView client when needed
OneView has the time to expire an active session
defined through an idle time. This idle time is
24 hours by default. The previous client,
python-oneviewclient, used by the hardware type
OneView had a reauthentication mechanism that would
ensure the new session was created when the old
one expired. With the migration to python-hpOneView,
the mechanism to reauthenticate does not exist.

This solution creates a new instance of the client when
it is needed.

Change-Id: Id112e84f7d050d818e3cd9bf02cd9cfe4716ef32
Closes-Bug: 1693788