Add missing job to gate
The job ironic-tempest-dsvm-ipa-wholedisk-bios-agent_ipmitool-tinyipa
is done in the check queue, but wasn't being done in gate. This adds
it to gate.

To make it easier to detect mismatches between the jobs in check vs
gate, the jobs are reordered so that they are in the same order in
check and gate, and the non-voting ones are grouped together.

Change-Id: Ic3d2571394be0e2b1eb73c044e1aea0ecdd3c864