Update Zuul 'gate' job
Have the Zuul 'gate' job
'ironic-tempest-dsvm-ipa-wholedisk-bios-agent_ipmitool-tinyipa' run on
master. This job runs in the Zuul 'check' job for master, so it should
also run in the 'gate' job. Although there are other jobs in 'check'
that aren't in 'gate', the others are non-voting. This one was in gate
prior to the migration to zuul v3. This fixes it so it is in gate again.

Change-Id: I9120dcf0b910759f60a4915c84388653922292e8