zuul.d/projects.yaml: Sort the job list
Sort the job list to make it easier to compare the 'check' jobs and
the 'gate' jobs.

Put the non-voting jobs in their own section.

Remove the 'voting: false' part from the experimental jobs as it is
unneeded.  Experimental jobs do not have any voting powers. So it is
redundant to mark them as 'voting: false'

Change-Id: I6ad8797c6973e04ec6835cfbab7b74eb0bd85220