Remove experimental jobs and sort job definitions in project.yaml
Remove the experimental jobs as they are not used in stable/pike

Remove the non-voting job:
tempest-dsvm-ironic-ipa-partition-pxe_ipmitool-tinyipa-python3 as we
did not get Python 3 support working in the Pike cycle.

Sort the job definitions in in project.yaml to make it easier to
compare and view.

Move non-voting jobs to end of list.

Delete unused playbooks due to deleting experimental jobs and the
Python 3 job.

Change-Id: Ia5c004bfde752f17d02ed784bd34cd9a60b6ff71