Don't collect logs from powered off nodes
The agent deploy mixin tries to collect logs from the deployment ramdisk
after powering off a node. This is clearly not going to work, as it
requires a call to the IPA API.

This change avoids collecting logs from the ramdisk in failure scenarios
where the node is powered off.

Change-Id: I353e73e3612ece893276c683ce469f8561e9ebf9
Closes-Bug: #1732939