Fix issue with double mocking of utils.execute functions
An issue was discovered if we mock an already mock-ed function. This
was happening in our execute() detection code. Change it to not use a
mock and instead a function.

Fix a failing unit test. This was due to the fact that using
'autospec=True' causes assert_not_called() to not work. This is a
known bug in the mock code[1].

Add test cases to ensure it works as expected.

[1] https://bugs.python.org/issue28380

Change-Id: Ifa18df61d17750f9a37118a2ac8aaf13005a5159