Run in superconductor cellsv2 mode for non-grenade jobs
Added in change I2bbdac5cd431e72517cb1c72bb925660ec8e868b,
ironic needed to run the multi-node jobs in singleconductor
mode because they rely on reschedules working in nova, which
doesn't work in superconductor mode where the top-level conductor
and scheduler are on a different message queue from the compute
service and cell conductor.

With the alternate hosts series in nova in Queens, we should have
the reschedule problem fixed such that when originally scheduling,
we not only get a target host but also a list of alternates for
retries within the cell. This allows us to run ironic multi-node
CI jobs (and all ironic non-upgrade related jobs for that matter)
in the default superconductor mode which is the recommended way
to run nova with cellsv2.

Depends-On: Iae904afb6cb4fcea8bb27741d774ffbe986a5fb4
Depends-On: Ie599968d9e7537e551fe6d9deb63a91b256b1e11
Change-Id: I755243abed054e6b6556a9a9498e565042293262