Check for PXE-enabled ports when creating neutron ports
The function "add_ports_to_network" will create neutron ports for each
PXE-enabled port on task.node to boot the ramdisk. But there is no
check to see if there are any PXE-enabled ports. If there aren't, no
neutron ports will be created and things will fail further down the
pipeline. To fail sooner with an appropriate message, we raise an
exception if there are no PXE-enabled ports for the node.

Story: 2001811
Task: 12546
Co-Authored-By: Ruby Loo <rloo@oath.com>
Change-Id: I96c21e7d842ad929161a0d298a342fdf4a1275e2